requirements.txt
bidict==0.23.1
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
coverage==7.9.1
docopt==0.6.2
Flask==3.1.1
Flask-SocketIO==5.5.1
Flask-Testing==0.8.1
h11==0.16.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
packaging==25.0
pipdeptree==2.26.1
pipreqs==0.4.13
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
pytest-cov==6.2.1
pytest-mock==3.14.1
python-engineio==4.12.2
python-socketio==5.13.0
requests==2.32.4
simple-websocket==1.1.0
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.2.0
yarg==0.1.10
